+++ crates/cli-tools/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Major
 
+- Remove `Option` from `action::RustAppletBuild::profile`
+- Remove `dir` argument from `action::RustApplet{Build,Test}::run()` (replaced with the `crate_dir`
+  field)
+- Rename `action::RustAppletBuild::output` to `output_dir` and remove `Option`
 - Change `action::{Transfer,RustApplet{New,Build,Test}}` to consume `self`
 - Remove `Default` for `action::RustAppletBuild` (implement `clap::Parser` instead)
 - Add `action` feature to gate the `action` module
@@ -11,6 +15,7 @@
 
 ### Minor
 
+- Add `action::RustApplet{Build,Test}::crate_dir`
 - Add `action::PlatformInfo` to print platform serial and version
 - Add `cmd::spawn()` for more control on command execution
 - Add `fs::remove_dir_all()` to remove a directory recursively